+++ test/regression/StaleNonceMetadataOverwrite.t.sol
@@ -17,6 +17,7 @@ import {IJBSuckerRegistry} from "../../src/interfaces/IJBSuckerRegistry.sol";
 import {JBInboxTreeRoot} from "../../src/structs/JBInboxTreeRoot.sol";
 import {JBMessageRoot} from "../../src/structs/JBMessageRoot.sol";
 import {ERC20Mock} from "../mocks/ERC20Mock.sol";
+import {JBConversionRate} from "../../src/structs/JBConversionRate.sol";
 
 /// @notice Harness exposing JBSwapCCIPSucker internals for testing stale nonce fix.
 contract StaleNonceTestHarness is JBSwapCCIPSucker {
@@ -47,7 +48,7 @@ contract StaleNonceTestHarness is JBSwapCCIPSucker {
         view
         returns (uint256 leafTotal, uint256 localTotal)
     {
-        ConversionRate storage rate = _conversionRateOf[token][nonce];
+        JBConversionRate storage rate = _conversionRateOf[token][nonce];
         return (rate.leafTotal, rate.localTotal);
     }